create session onController hook
