+++ packages/keychain/src/components/v2/Auth/Login.tsx
@@ -15,8 +15,6 @@ import base64url from "base64url";
 import { AccountQuery } from "generated/graphql";
 import { useClearField, useSubmitType, useUsername } from "./hooks";
 import { validateUsername } from "./validate";
-import { Status } from "utils/account";
-import { ResponseCodes, Error } from "@cartridge/controller";
 
 export function Login({
   fullPage = false,
@@ -58,53 +56,11 @@ export function Login({
         const controller = new Controller(keypair, address, credentialId);
 
         if (onController) {
-          onController(controller);
+          await onController(controller);
         }
 
         if (onComplete) {
           onComplete();
-        }
-
-        // check session
-        if (context) {
-          const account = controller.account(
-            (context as any).transactionsDetail?.chainId ?? chainId,
-          );
-          const sesh = controller.session(context.origin);
-
-          if (!sesh) {
-            if (account.status === Status.COUNTERFACTUAL) {
-              // TODO: Deploy?
-              context.resolve({
-                code: ResponseCodes.SUCCESS,
-                address: controller.address,
-                policies: context.policies,
-              } as any);
-              return;
-            }
-
-            // This device needs to be registered, so do a webauthn signature request
-            // for the register transaction during the connect flow.
-            if (account.status === Status.DEPLOYED) {
-              try {
-                await account.register();
-              } catch (e) {
-                context.resolve({
-                  code: ResponseCodes.CANCELED,
-                  message: "Canceled",
-                } as Error);
-                return;
-              }
-            }
-
-            controller.approve(context.origin, context.policies, "");
-
-            context.resolve({
-              code: ResponseCodes.SUCCESS,
-              address: controller.address,
-              policies: context.policies,
-            } as any);
-          }
         }
       } catch (err) {
         setIsLoggingIn(false);
@@ -116,15 +72,14 @@ export function Login({
         });
       }
     },
-    [account, log, onComplete, onController, chainId, context],
+    [account, log, onComplete, onController],
   );
 
   return (
-    <Container fullPage={fullPage}>
+    <Container fullPage={fullPage} chainId={chainId}>
       <Formik initialValues={{ username: prefilledName }} onSubmit={onSubmit}>
         <Form
           context={context}
-          onController={onController}
           onSignup={onSignup}
           setAccount={setAccount}
           isLoggingIn={isLoggingIn}
@@ -136,11 +91,10 @@ export function Login({
 
 function Form({
   context,
-  // onController,
   onSignup,
   setAccount,
   isLoggingIn,
-}: Pick<LoginProps, "context" | "onController" | "onSignup"> & {
+}: Pick<LoginProps, "context" | "onSignup"> & {
   setAccount: (account: AccountQuery["account"]) => void;
   isLoggingIn: boolean;
 }) {
@@ -177,7 +131,7 @@ function Form({
         </FormikField>
       </VStack>
 
-      <PortalFooter origin={context.origin} policies={context.policies}>
+      <PortalFooter origin={context?.origin} policies={context?.policies}>
         <VStack
           w="full"
           alignItems="flex"
